extracted FeedVideo component, removed unnecessary creation of IntersectionObserver
